Moves the Executor dyn Trait to instruction_processor.rs
